Replace `DrugViewHolder` with a custom widget
